jagged candidate array documentation improvments
