+++ coffea/analysis_objects/JaggedTLorentzVectorArray.py
@@ -1,0 +1,4 @@
+import uproot_methods
+from ..util import awkward
+
+JaggedTLorentzVectorArray = awkward.Methods.mixin(uproot_methods.classes.TLorentzVector.ArrayMethods, awkward.JaggedArray)
