+++ coffea/analysis_objects/__init__.py
@@ -1,13 +1,11 @@
 """Physics candidate wrapper classes for awkward arrays
-
-Details...
 """
-from .JaggedCandidateArray import (
-    JaggedCandidateArray,
-    JaggedTLorentzVectorArray,
-)
+from .JaggedCandidateArray import JaggedCandidateArray
+from .JaggedCandidateMethods import JaggedCandidateMethods
+from .JaggedTLorentzVectorArray import JaggedTLorentzVectorArray
 
 __all__ = [
     'JaggedCandidateArray',
+    'JaggedCandidateMethods',
     'JaggedTLorentzVectorArray'
 ]
